LayoutTests/media/media-source/media-source-ended-expected.txt


EVENT(update)
EVENT(update)
EVENT(ended)
EXPECTED (video.currentTime == '1') OK
END OF TEST
